soc: rtl87x2g: customize k_busy_wait api
The native k_busy_wait in Zephyr is implemented based on polling the
systick counter for delays. In practice, on the KM4, this implementation
results in inaccurate delays. Therefore, the ARCH_HAS_CUSTOM_BUSY_WAIT
is enabled to use the platform_delay_us api from RTK to customize
the implementation of k_busy_wait.

Signed-off-by: Zhiyuan Tang <zhiyuan_tang@realsil.com.cn>